+++ src/dipdup/cli.py
@@ -42,6 +42,14 @@ NO_CONFIG_CMDS = {
     'config',
 }
 
+# NOTE: commands we reuse from aerich for database migration
+AERICH_CMDS = {
+    'history',
+    'heads',
+    'migrate',
+    'upgrade',
+    'downgrade',
+}
 
 _logger = logging.getLogger(__name__)
 _click_wrap_text = click.formatting.wrap_text
@@ -149,6 +157,14 @@ WrappedCommandT = TypeVar('WrappedCommandT', bound=Callable[..., Coroutine[Any, 
 class CLIContext:
     config_paths: list[str]
     config: 'DipDupConfig'
+
+    # NOTE: we have to implement __getitem__ and __setitem__ because aerich CLI
+    #  commands (`AERICH_CMDS`) access the context object via `ctx.obj["command"]`
+    def __getitem__(self, item: str) -> Any:
+        return getattr(self, item)
+
+    def __setitem__(self, key: str, value: Any) -> None:
+        setattr(self, key, value)
 
 
 def _cli_wrapper(fn: WrappedCommandT) -> WrappedCommandT:
@@ -465,6 +481,7 @@ async def migrate(ctx: click.Context, dry_run: bool) -> None:
         unsafe=True,
     )
     config.initialize()
+
     ctx.obj = CLIContext(
         config_paths=ctx.parent.params['config'],
         config=config,
@@ -627,7 +644,60 @@ async def hasura_configure(ctx: click.Context, force: bool) -> None:
 @_cli_wrapper
 async def schema(ctx: click.Context) -> None:
     """Commands to manage database schema."""
-    pass
+    if '--help' in sys.argv:
+        return
+
+    config: DipDupConfig = ctx.obj.config
+
+    if ctx.invoked_subcommand in AERICH_CMDS:
+        from dipdup.package import DipDupPackage
+
+        migrations_dir = DipDupPackage(config.package_path).migrations
+
+        if not migrations_dir.exists():
+            from dipdup.exceptions import ProjectPackageError
+
+            raise ProjectPackageError(
+                f"""Database migrations are not initialized at {migrations_dir}.
+              Run `dipdup schema init` or `dipdup run` to the run the indexer and it'll be initialized automatically."""
+            )
+
+        from dipdup.aerich import create_aerich_command
+
+        aerich_command = await create_aerich_command(config.database.connection_string, config.package, migrations_dir)
+        await aerich_command.init()
+
+        ctx.obj['command'] = aerich_command
+
+
+# Wrapper function to approve schema after upgrade and downgrade
+# Since upgrade and downgrade are meant to be intentionally run by the user, it'll be confusing for the user to run approve just after them
+def _approve_schema_after(command: click.Command) -> click.Command:
+    @click.pass_context
+    def wrapper(ctx: click.Context, /, *args: Any, **kwargs: Any) -> None:
+        ctx.invoke(command, *args, **kwargs)
+        # TODO: don't call approve if no upgrades/downgrades happened
+        ctx.invoke(schema_approve)
+
+    return click.Command(
+        name=command.name,
+        callback=wrapper,
+        params=command.params,
+        help=command.help,
+        short_help=command.short_help,
+    )
+
+
+try:
+    from aerich.cli import cli as aerich_cli
+
+    schema.add_command(aerich_cli.commands['history'])
+    schema.add_command(aerich_cli.commands['heads'])
+    schema.add_command(aerich_cli.commands['migrate'])
+    schema.add_command(_approve_schema_after(aerich_cli.commands['upgrade']))
+    schema.add_command(_approve_schema_after(aerich_cli.commands['downgrade']))
+except ImportError:
+    _logger.debug('aerich is not installed, skipping database migration commands')
 
 
 @schema.command(name='approve')
@@ -712,6 +782,9 @@ async def schema_wipe(ctx: click.Context, immune: bool, force: bool) -> None:
     from dipdup.database import get_connection
     from dipdup.database import tortoise_wrapper
     from dipdup.database import wipe_schema
+    from dipdup.package import DipDupPackage
+
+    migrations_dir = DipDupPackage(config.package_path).migrations
 
     async with tortoise_wrapper(
         url=url,
@@ -729,6 +802,7 @@ async def schema_wipe(ctx: click.Context, immune: bool, force: bool) -> None:
                 else config.database.schema_name
             ),
             immune_tables=immune_tables,
+            migrations_dir=migrations_dir,
         )
 
     _logger.info('Schema wiped')
@@ -758,6 +832,7 @@ async def schema_init(ctx: click.Context) -> None:
         await dipdup._set_up_hooks()
         await dipdup._create_datasources()
         await dipdup._initialize_schema()
+        await dipdup._initialize_migrations()
 
         # NOTE: It's not necessary a reindex, but it's safe to execute built-in scripts to (re)create views.
         conn = get_connection()
@@ -854,6 +929,7 @@ async def new(
     _logger.info('Initializing project')
     config = DipDupConfig.load([Path(answers['package'])])
     config.initialize()
+
     ctx.obj = CLIContext(
         config_paths=[Path(answers['package']).joinpath(ROOT_CONFIG).as_posix()],
         config=config,